+++ src/backend/ci/core/auth/biz-auth-rbac/src/main/kotlin/com/tencent/devops/auth/service/RbacPermissionResourceMemberService.kt
@@ -1,0 +1,136 @@
+package com.tencent.devops.auth.service
+
+import com.tencent.bk.sdk.iam.constants.ManagerScopesEnum
+import com.tencent.bk.sdk.iam.dto.PageInfoDTO
+import com.tencent.bk.sdk.iam.dto.V2PageInfoDTO
+import com.tencent.bk.sdk.iam.dto.manager.V2ManagerRoleGroupInfo
+import com.tencent.bk.sdk.iam.dto.manager.dto.SearchGroupDTO
+import com.tencent.bk.sdk.iam.service.v2.V2ManagerService
+import com.tencent.devops.auth.dao.AuthResourceGroupDao
+import com.tencent.devops.auth.service.iam.PermissionResourceMemberService
+import com.tencent.devops.common.auth.api.AuthResourceType
+import com.tencent.devops.common.auth.api.pojo.BkAuthGroup
+import com.tencent.devops.common.auth.api.pojo.BkAuthGroupAndUserList
+import org.jooq.DSLContext
+import org.slf4j.LoggerFactory
+
+class RbacPermissionResourceMemberService constructor(
+    private val authResourceService: AuthResourceService,
+    private val iamV2ManagerService: V2ManagerService,
+    private val permissionGradeManagerService: PermissionGradeManagerService,
+    private val authResourceGroupDao: AuthResourceGroupDao,
+    private val dslContext: DSLContext,
+) : PermissionResourceMemberService {
+    override fun getResourceGroupMembers(
+        projectCode: String,
+        resourceType: String,
+        resourceCode: String,
+        group: String?
+    ): List<String> {
+        return when (group) {
+            // 新的rbac版本中，没有ci管理员组，不可以调用此接口来获取ci管理员组的成员
+            BkAuthGroup.CIADMIN.value, BkAuthGroup.CI_MANAGER.value -> emptyList()
+            // 获取特定资源下全部成员
+            null -> {
+                getResourceGroupAndMembers(
+                    projectCode = projectCode,
+                    resourceType = resourceType,
+                    resourceCode = resourceCode
+                ).flatMap { it.userIdList }.distinct()
+            }
+            // 获取特定资源下特定用户组成员
+            else -> {
+                val dbGroupInfo = authResourceGroupDao.get(
+                    dslContext = dslContext,
+                    projectCode = projectCode,
+                    resourceType = resourceType,
+                    resourceCode = resourceCode,
+                    groupCode = group
+                ) ?: return emptyList()
+                val groupInfo = getResourceGroupAndMembers(
+                    projectCode = projectCode,
+                    resourceType = resourceType,
+                    resourceCode = resourceCode
+                ).find { it.roleId == dbGroupInfo.relationId.toInt() }
+                groupInfo?.userIdList ?: emptyList()
+            }
+        }
+    }
+
+    override fun getResourceGroupAndMembers(
+        projectCode: String,
+        resourceType: String,
+        resourceCode: String
+    ): List<BkAuthGroupAndUserList> {
+        // 1、获取管理员id
+        val managerId = authResourceService.get(
+            projectCode = projectCode,
+            resourceType = resourceType,
+            resourceCode = resourceCode
+        ).relationId
+        // 2、获取分级管理员下所有的用户组
+        val groupInfoList = getGroupInfoList(
+            resourceType = resourceType,
+            managerId = managerId
+        )
+        logger.info(
+            "[RBAC-IAM] getResourceGroupAndMembers: projectCode = $projectCode |" +
+                " managerId = $managerId | groupInfoList: $groupInfoList"
+        )
+        // 3、获取组成员
+        return groupInfoList.map { getUsersUnderGroup(groupInfo = it) }
+    }
+
+    private fun getGroupInfoList(
+        resourceType: String,
+        managerId: String
+    ): List<V2ManagerRoleGroupInfo> {
+        return if (resourceType == AuthResourceType.PROJECT.value) {
+            val searchGroupDTO = SearchGroupDTO.builder().inherit(false).build()
+            permissionGradeManagerService.listGroup(
+                gradeManagerId = managerId,
+                searchGroupDTO = searchGroupDTO,
+                page = 1,
+                pageSize = 1000
+            )
+        } else {
+            val v2PageInfoDTO = V2PageInfoDTO().apply {
+                pageSize = 1000
+                page = 1
+            }
+            iamV2ManagerService.getSubsetManagerRoleGroup(
+                managerId.toInt(),
+                v2PageInfoDTO
+            ).results
+        }
+    }
+
+    private fun getUsersUnderGroup(groupInfo: V2ManagerRoleGroupInfo): BkAuthGroupAndUserList {
+        val pageInfoDTO = PageInfoDTO().apply {
+            limit = 1000
+            offset = 0
+        }
+        val groupMemberInfoList = iamV2ManagerService.getRoleGroupMemberV2(groupInfo.id, pageInfoDTO).results
+        logger.info(
+            "[RBAC-IAM] getUsersUnderGroup ,groupId: ${groupInfo.id} " +
+                "| groupMemberInfoList: $groupMemberInfoList"
+        )
+        val members = mutableListOf<String>()
+        groupMemberInfoList.forEach { memberInfo ->
+            if (memberInfo.type == ManagerScopesEnum.getType(ManagerScopesEnum.USER)) {
+                members.add(memberInfo.id)
+            }
+        }
+        return BkAuthGroupAndUserList(
+            displayName = groupInfo.name,
+            roleId = groupInfo.id,
+            roleName = groupInfo.name,
+            userIdList = members.toSet().toList(),
+            type = ""
+        )
+    }
+
+    companion object {
+        private val logger = LoggerFactory.getLogger(RbacPermissionResourceMemberService::class.java)
+    }
+}
