+++ src/backend/ci/core/auth/biz-auth/src/main/kotlin/com/tencent/devops/auth/service/iam/PermissionProjectService.kt
@@ -32,22 +32,12 @@ import com.tencent.devops.common.auth.api.pojo.BkAuthGroup
 import com.tencent.devops.common.auth.api.pojo.BkAuthGroupAndUserList
 
 interface PermissionProjectService {
-
-    /**
-     * 获取特定资源下的用户组用户
-     * */
     fun getProjectUsers(
         projectCode: String,
-        resourceType: String,
-        resourceCode: String,
         group: String?
     ): List<String>
 
-    fun getProjectGroupAndUserList(
-        projectCode: String,
-        resourceType: String,
-        resourceCode: String
-    ): List<BkAuthGroupAndUserList>
+    fun getProjectGroupAndUserList(projectCode: String): List<BkAuthGroupAndUserList>
 
     fun getUserProjects(userId: String): List<String>
