+++ src/backend/ci/core/auth/biz-auth/src/main/kotlin/com/tencent/devops/auth/service/sample/SamplePermissionResourceMemberService.kt
@@ -1,0 +1,23 @@
+package com.tencent.devops.auth.service.sample
+
+import com.tencent.devops.auth.service.iam.PermissionResourceMemberService
+import com.tencent.devops.common.auth.api.pojo.BkAuthGroupAndUserList
+
+class SamplePermissionResourceMemberService : PermissionResourceMemberService {
+    override fun getResourceGroupMembers(
+        projectCode: String,
+        resourceType: String,
+        resourceCode: String,
+        group: String?
+    ): List<String> {
+        return emptyList()
+    }
+
+    override fun getResourceGroupAndMembers(
+        projectCode: String,
+        resourceType: String,
+        resourceCode: String
+    ): List<BkAuthGroupAndUserList> {
+        return emptyList()
+    }
+}
